+++ init.lua
@@ -1,38 +1,15 @@
---[[
-------------------------------------------------------------------------------
-Simple Tiled Implementation is licensed under the MIT Open Source License.
-(http://www.opensource.org/licenses/mit-license.html)
-------------------------------------------------------------------------------
-
-Copyright (c) 2015 Landon Manning - LManning17@gmail.com - LandonManning.com
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-]]--
+-- @author Landon Manning
+-- @copyright 2015
+-- @license MIT/X11
 
 local STI = {
 	_LICENSE     = "STI is distributed under the terms of the MIT license. See LICENSE.md.",
 	_URL         = "https://github.com/karai17/Simple-Tiled-Implementation",
-	_VERSION     = "0.13.1.0",
+	_VERSION     = "0.13.1.1",
 	_DESCRIPTION = "Simple Tiled Implementation is a Tiled Map Editor library designed for the *awesome* LÖVE framework."
 }
 
-local path = ... .. "." -- lol
+local path = (...):gsub('%.init$', '') .. "."
 local Map  = require(path .. "map")
 local framework
 
@@ -42,10 +19,14 @@ else
 	framework = require(path .. "framework.lua")
 end
 
-function STI.new(map)
+--- Instance a new map.
+-- @param path Path to the map file.
+-- @param plugins A list of plugins to load.
+-- @return table The loaded Map.
+function STI.new(map, plugins)
 	-- Check for valid map type
-	local ext = map:sub(-3, -1)
-	assert(ext == "lua", string.format(
+	local ext = map:sub(-4, -1)
+	assert(ext == ".lua", string.format(
 		"Invalid file type: %s. File must be of type: lua.",
 		ext
 	))
@@ -61,7 +42,7 @@ function STI.new(map)
 	setfenv(map, {})
 	map = setmetatable(map(), {__index = Map})
 
-	map:init(path, framework)
+	map:init(path, framework, plugins)
 
 	return map
 end