+++ map.lua
@@ -1,4 +1,6 @@
-local Map = {}
+local path       = (...):gsub('%.[^%.]+$', '') .. "."
+local pluginPath = string.gsub(path, "[.]", "/") .. "plugins/"
+local Map        = {}
 local framework
 
 -- https://github.com/stevedonovan/Penlight/blob/master/lua/pl/path.lua#L286
@@ -20,8 +22,18 @@ local function formatPath(path)
 	return path
 end
 
-function Map:init(path, fw)
-	framework          = fw
+--- Instance a new map.
+-- @param path Path to the map file.
+-- @param fw The framework being used.
+-- @param plugins A list of plugins to load.
+-- @return nil
+function Map:init(path, fw, plugins)
+	framework = fw
+
+	if type(plugins) == "table" then
+		self:loadPlugins(plugins)
+	end
+
 	self.canvas        = framework:newCanvas()
 	self.objects       = {}
 	self.tiles         = {}
@@ -49,6 +61,31 @@ function Map:init(path, fw)
 	end
 end
 
+--- Load plugins.
+-- @param plugins A list of plugins to load.
+-- @return nil
+function Map:loadPlugins(plugins)
+	for _, plugin in ipairs(plugins) do
+		local p = pluginPath .. plugin .. ".lua"
+		if framework.isFile(p) then
+			local file = framework.load(p)()(framework)
+			for k, func in pairs(file) do
+				if not self[k] then
+					self[k] = func
+				end
+			end
+		end
+	end
+
+	--for _, file in ipairs(framework.getDirectoryItems(pluginPath)) do
+	--end
+end
+
+--- Create Tiles.
+-- @param index Index of the Tileset.
+-- @param tileset Tileset data.
+-- @param gid First Global ID in Tileset.
+-- @return number Next Tileset's first Global ID
 function Map:setTiles(index, tileset, gid)
 	local function getTiles(i, t, m, s)
 		i = i - m
@@ -127,6 +164,10 @@ function Map:setTiles(index, tileset, gid)
 	return gid
 end
 
+--- Create Layers.
+-- @param layer Layer data.
+-- @param path (Optional) Path to an Image Layer's image.
+-- @return nil
 function Map:setLayer(layer, path)
 	layer.x      = layer.x or 0
 	layer.y      = layer.y or 0
@@ -155,6 +196,9 @@ function Map:setLayer(layer, path)
 	self.layers[layer.name] = layer
 end
 
+--- Add Tiles to Tile Layer.
+-- @param layer The Tile Layer.
+-- @return nil
 function Map:setTileData(layer)
 	local i   = 1
 	local map = {}
@@ -175,12 +219,18 @@ function Map:setTileData(layer)
 	layer.data = map
 end
 
+--- Add Objects to Layer.
+-- @param layer The Object Layer.
+-- @return nil
 function Map:setObjectData(layer)
 	for _, object in ipairs(layer.objects) do
 		self.objects[object.id] = object
 	end
 end
 
+--- Correct position and orientation of Objects in an Object Layer.
+-- @param layer The Object Layer.
+-- @return nil
 function Map:setObjectCoordinates(layer)
 	local function convertEllipseToPolygon(x, y, w, h, max_segments)
 		local function calc_segments(segments)
@@ -311,6 +361,9 @@ function Map:setObjectCoordinates(layer)
 	end
 end
 
+--- Batch Tiles in Tile Layer for improved draw speed.
+-- @param layer The Tile Layer.
+-- @return nil
 function Map:setSpriteBatches(layer)
 	local newBatch = framework.newSpriteBatch
 	local w        = framework.getWidth()
@@ -395,6 +448,9 @@ function Map:setSpriteBatches(layer)
 	layer.batches = batches
 end
 
+--- Batch Tiles in Object Layer for improved draw speed.
+-- @param layer The Object Layer.
+-- @return nil
 function Map:setObjectSpriteBatches(layer)
 	local newBatch = framework.newSpriteBatch
 	local tw       = self.tilewidth
@@ -429,6 +485,12 @@ function Map:setObjectSpriteBatches(layer)
 	layer.batches = batches
 end
 
+--- Only draw what is visible on screen for improved draw speed.
+-- @param tx Translate X axis (in pixels).
+-- @param ty Translate Y axis (in pixels).
+-- @param w Width of screen (in pixels).
+-- @param h Height of screen (in pixels).
+-- @return nil
 function Map:setDrawRange(tx, ty, w, h)
 	local tw, th = self.tilewidth, self.tileheight
 	local sx, sy, ex, ey
@@ -458,6 +520,10 @@ function Map:setDrawRange(tx, ty, w, h)
 	}
 end
 
+--- Create a Custom Layer to place userdata in (such as player sprites).
+-- @param name Name of Custom Layer.
+-- @param index Draw order within Layer stack.
+-- @return table Custom Layer
 function Map:addCustomLayer(name, index)
 	local layer = {
       type       = "customlayer",
@@ -476,6 +542,9 @@ function Map:addCustomLayer(name, index)
 	return layer
 end
 
+--- Convert another Layer into a Custom Layer.
+-- @param index Index or name of Layer to convert.
+-- @return table Custom Layer
 function Map:convertToCustomLayer(index)
 	local layer = assert(self.layers[index], "Layer not found: " .. index)
 
@@ -495,6 +564,9 @@ function Map:convertToCustomLayer(index)
 	return layer
 end
 
+--- Remove a Layer from the Layer stack.
+-- @param index Index or name of Layer to convert.
+-- @return nil
 function Map:removeLayer(index)
 	local layer = assert(self.layers[index], "Layer not found: " .. index)
 
@@ -513,6 +585,9 @@ function Map:removeLayer(index)
 	end
 end
 
+--- Animate Tiles and update every Layer.
+-- @param dt Delta Time.
+-- @return nil
 function Map:update(dt)
 	for gid, tile in pairs( self.tiles ) do
 		local update, t
@@ -544,6 +619,8 @@ function Map:update(dt)
 	end
 end
 
+--- Draw every Layer.
+-- @return nil
 function Map:draw()
 	local current_canvas = framework.getCanvas()
 	framework.setCanvas(self.canvas)
@@ -567,12 +644,18 @@ function Map:draw()
 	framework.pop()
 end
 
+--- Draw an individual Layer.
+-- @param layer The Layer to draw.
+-- @return nil
 function Map:drawLayer(layer)
 	framework.setColor(255, 255, 255, 255 * layer.opacity)
 	layer:draw()
 	framework.setColor(255, 255, 255, 255)
 end
 
+--- Default draw function for Tile Layers.
+-- @param layer The Tile Layer to draw.
+-- @return nil
 function Map:drawTileLayer(layer)
 	if type(layer) == "string" or type(layer) == "number" then
 		layer = self.layers[layer]
@@ -604,6 +687,9 @@ function Map:drawTileLayer(layer)
 	end
 end
 
+--- Default draw function for Object Layers.
+-- @param layer The Object Layer to draw.
+-- @return nil
 function Map:drawObjectLayer(layer)
 	if type(layer) == "string" or type(layer) == "number" then
 		layer = self.layers[layer]
@@ -678,6 +764,9 @@ function Map:drawObjectLayer(layer)
 	end
 end
 
+--- Default draw function for Image Layers.
+-- @param layer The Image Layer to draw.
+-- @return nil
 function Map:drawImageLayer(layer)
 	if type(layer) == "string" or type(layer) == "number" then
 		layer = self.layers[layer]
@@ -690,10 +779,17 @@ function Map:drawImageLayer(layer)
 	end
 end
 
+--- Resize the drawable area of the Map.
+-- @param w The new width of the drawable area (in pixels).
+-- @param h The new Height of the drawable area (in pixels).
+-- @return nil
 function Map:resize(w, h)
 	self.canvas = framework:newCanvas(w, h)
 end
 
+--- Create flipped or rotated Tiles based on bitop flags.
+-- @param gid The flagged Global ID.
+-- @return table Flipped Tile
 function Map:setFlippedGID(gid)
 	local bit31   = 2147483648
 	local bit30   = 1073741824
@@ -765,6 +861,9 @@ function Map:setFlippedGID(gid)
 	return self.tiles[gid]
 end
 
+--- Get custom properties from Layer.
+-- @param layer The Layer.
+-- @return table List of properties
 function Map:getLayerProperties(layer)
 	local l = self.layers[layer]
 
@@ -773,6 +872,11 @@ function Map:getLayerProperties(layer)
 	return l.properties
 end
 
+--- Get custom properties from Tile.
+-- @param layer The Layer that the Tile belongs to.
+-- @param x The X axis location of the Tile (in tiles).
+-- @param y The Y axis location of the Tile (in tiles).
+-- @return table List of properties
 function Map:getTileProperties(layer, x, y)
 	local tile = self.layers[layer].data[y][x]
 
@@ -781,6 +885,10 @@ function Map:getTileProperties(layer, x, y)
 	return tile.properties
 end
 
+--- Get custom properties from Object.
+-- @param layer The Layer that the Object belongs to.
+-- @param object The index or name of the Object.
+-- @return table List of properties
 function Map:getObjectProperties(layer, object)
 	local o = self.layers[layer].objects
 
@@ -800,6 +908,11 @@ function Map:getObjectProperties(layer, object)
 	return o.properties
 end
 
+--- Project isometric position to orthoganal position.
+-- @param x The X axis location of the point (in pixels).
+-- @param y The Y axis location of the point (in pixels).
+-- @return number The X axis location of the point (in pixels).
+-- @return number The Y axis location of the point (in pixels).
 function Map:convertIsometricToScreen(x, y)
 	local mw = self.width
 	local tw = self.tilewidth
@@ -811,6 +924,11 @@ function Map:convertIsometricToScreen(x, y)
 	return sx, sy
 end
 
+--- Project orthoganal position to isometric position.
+-- @param x The X axis location of the point (in pixels).
+-- @param y The Y axis location of the point (in pixels).
+-- @return number The X axis location of the point (in pixels).
+-- @return number The Y axis location of the point (in pixels).
 function Map:convertScreenToIsometric(x, y)
 	local mw = self.width
 	local mh = self.height
@@ -824,6 +942,11 @@ function Map:convertScreenToIsometric(x, y)
 	return tx, ty
 end
 
+--- Convert orthoganal tile space to screen space.
+-- @param x The X axis location of the point (in tiles).
+-- @param y The Y axis location of the point (in tiles).
+-- @return number The X axis location of the point (in pixels).
+-- @return number The Y axis location of the point (in pixels).
 function Map:convertTileToScreen(x, y)
 	local tw = self.tilewidth
 	local th = self.tileheight
@@ -833,6 +956,11 @@ function Map:convertTileToScreen(x, y)
 	return sx, sy
 end
 
+--- Convert orthoganal screen space to tile space.
+-- @param x The X axis location of the point (in pixels).
+-- @param y The Y axis location of the point (in pixels).
+-- @return number The X axis location of the point (in tiles).
+-- @return number The Y axis location of the point (in tiles).
 function Map:convertScreenToTile(x, y)
 	local tw = self.tilewidth
 	local th = self.tileheight
@@ -842,6 +970,11 @@ function Map:convertScreenToTile(x, y)
 	return tx, ty
 end
 
+--- Convert isometric tile space to screen space.
+-- @param x The X axis location of the point (in tiles).
+-- @param y The Y axis location of the point (in tiles).
+-- @return number The X axis location of the point (in pixels).
+-- @return number The Y axis location of the point (in pixels).
 function Map:convertIsometricTileToScreen(x, y)
 	local mw = self.width
 	local tw = self.tilewidth
@@ -853,6 +986,11 @@ function Map:convertIsometricTileToScreen(x, y)
 	return sx, sy
 end
 
+--- Convert isometric screen space to tile space.
+-- @param x The X axis location of the point (in pixels).
+-- @param y The Y axis location of the point (in pixels).
+-- @return number The X axis location of the point (in tiles).
+-- @return number The Y axis location of the point (in tiles).
 function Map:convertScreenToIsometricTile(x, y)
 	local mw = self.width
 	local tw = self.tilewidth
@@ -864,6 +1002,11 @@ function Map:convertScreenToIsometricTile(x, y)
 	return tx, ty
 end
 
+--- Convert staggered isometric tile space to screen space.
+-- @param x The X axis location of the point (in tiles).
+-- @param y The Y axis location of the point (in tiles).
+-- @return number The X axis location of the point (in pixels).
+-- @return number The Y axis location of the point (in pixels).
 function Map:convertStaggeredTileToScreen(x, y)
 	local tw = self.tilewidth
 	local th = self.tileheight
@@ -873,6 +1016,11 @@ function Map:convertStaggeredTileToScreen(x, y)
 	return sx, sy
 end
 
+--- Convert staggered isometric screen space to tile space.
+-- @param x The X axis location of the point (in pixels).
+-- @param y The Y axis location of the point (in pixels).
+-- @return number The X axis location of the point (in tiles).
+-- @return number The Y axis location of the point (in tiles).
 function Map:convertScreenToStaggeredTile(x, y)
 	local function topLeft(x, y)
 		if (math.ceil(y) % 2) then